add: adding more status and listener now is a separate file
